+++ examples/nuxt/nuxt.config.ts
@@ -1,6 +1,0 @@
-import "./env";
-
-// https://nuxt.com/docs/api/configuration/nuxt-config
-export default defineNuxtConfig({
-  // ...
-});