+++ examples/nuxt/server/api/hello.get.ts
@@ -1,7 +1,0 @@
-import { env } from "~~/env";
-
-export default defineEventHandler(() => {
-  return {
-    secret: `Server says ${env.SECRET}`,
-  };
-});