+++ packages/core/test/smoke-valibot.test.ts
@@ -1,14 +1,12 @@
-/// <reference types="bun" />
-import { describe, expect, spyOn, test } from "bun:test";
-import { expectTypeOf } from "expect-type";
-
 import * as v from "valibot";
-import { createEnv } from "../src";
+import { describe, expect, expectTypeOf, test, vi } from "vitest";
+import { createEnv } from "../src/index.ts";
+import { uploadthing } from "../src/presets-valibot.ts";
 
 function ignoreErrors(cb: () => void) {
   try {
     cb();
-  } catch (err) {
+  } catch {
     // ignore
   }
 }
@@ -231,15 +229,11 @@ describe("errors when validation fails", () => {
           FOO_BAR: "foo",
         },
         onValidationError: (issues) => {
-          const barError = issues.find(
-            (issue) => issue.path?.[0] === "BAR",
-          )?.message;
+          const barError = issues.find((issue) => issue.path?.[0] === "BAR")?.message;
           throw new Error(`Invalid variable BAR: ${barError}`);
         },
       }),
-    ).toThrow(
-      "Invalid variable BAR: Invalid type: Expected number but received NaN",
-    );
+    ).toThrow("Invalid variable BAR: Invalid type: Expected number but received NaN");
   });
 });
 
@@ -452,7 +446,7 @@ describe("extending presets", () => {
       }>
     >();
 
-    const consoleError = spyOn(console, "error");
+    const consoleError = vi.spyOn(console, "error");
     expect(() => lazyCreateEnv()).toThrow("Invalid environment variables");
     expect(consoleError.mock.calls[0]).toEqual([
       "❌ Invalid environment variables:",
@@ -712,8 +706,7 @@ describe("createFinalSchema", () => {
             if (env.SKIP_AUTH) return { SKIP_AUTH: true } as const;
             if (!env.EMAIL || !env.PASSWORD) {
               addIssue({
-                message:
-                  "EMAIL and PASSWORD are required if SKIP_AUTH is false",
+                message: "EMAIL and PASSWORD are required if SKIP_AUTH is false",
               });
               return NEVER;
             }
@@ -776,10 +769,7 @@ test("overriding preset env var", () => {
 
   const env = createEnv({
     server: {
-      PRESET_ENV: v.union([
-        v.pipe(v.string(), v.transform(Number)),
-        v.number(),
-      ]),
+      PRESET_ENV: v.union([v.pipe(v.string(), v.transform(Number)), v.number()]),
     },
     extends: [preset],
     runtimeEnv: { PRESET_ENV: 123 },
@@ -791,4 +781,25 @@ test("overriding preset env var", () => {
     }>
   >();
   expect(env.PRESET_ENV).toBe(123);
+});
+
+test("with built-in preset", () => {
+  process.env.UPLOADTHING_TOKEN = "token";
+  const env = createEnv({
+    server: {
+      FOO: v.string(),
+    },
+    extends: [uploadthing()],
+    runtimeEnv: { FOO: "bar" },
+  });
+
+  expectTypeOf(env).toEqualTypeOf<
+    Readonly<{
+      FOO: string;
+      UPLOADTHING_TOKEN: string;
+    }>
+  >();
+
+  expect(env.FOO).toBe("bar");
+  expect(env.UPLOADTHING_TOKEN).toBe("token");
 });