+++ packages/nuxt/src/presets-arktype.ts
@@ -1,0 +1,6 @@
+/**
+ * Presets for Arktype
+ * @see https://env.t3.gg/docs/customization#extending-presets
+ * @module
+ */
+export * from "@t3-oss/env-core/presets-arktype";